+++ lib/adhearsion/call_controller.rb
@@ -118,7 +118,7 @@ module Adhearsion
       call.async.register_controller self
       execute_callbacks :before_call
       run
-    rescue Call::Hangup
+    rescue Call::Hangup, Call::ExpiredError
       logger.info "Call was hung up while executing a controller"
     rescue SyntaxError, StandardError => e
       Events.trigger :exception, [e, logger]