+++ CHANGELOG.md
@@ -23,6 +23,7 @@
   * Bugfix: Fix controller block context on JRuby
   * Bugfix: Require only Bundler groups appropriate to application environment
   * Bugfix: Call event handlers now cannot crash the call, and no longer stop on return value
+  * Bugfix: Call controllers are gracefully terminated when a command is attempted against a dead call
 
 # [2.4.0](https://github.com/adhearsion/adhearsion/compare/v2.3.5...v2.4.0) - [2013-08-29](https://rubygems.org/gems/adhearsion/versions/2.4.0)
   * Deprecation: Ruby 1.9.2 support is deprecated and will be dropped in a future version of Adhearsion